refactor: [GNS-258] AccountDetail
